requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
astroid==4.0.4
backoff==2.2.1
bcrypt==5.0.0
black==26.1.0
blinker==1.9.0
build==1.4.0
cachelib==0.13.0
certifi==2026.2.25
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.3.1
dill==0.4.1
distro==1.9.0
durationpy==0.10
fastapi==0.134.0
filelock==3.20.0
Flask==3.0.3
Flask-Caching==2.3.1
Flask-Cors==4.0.1
flatbuffers==25.12.19
fsspec==2025.12.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
httptools==0.7.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
isort==8.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kubernetes==35.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.2
posthog==7.9.4
protobuf==6.33.5
pulsar-client==3.10.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pylint==4.0.5
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytokens==0.4.1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.14.0
torch==2.2.1+cpu
tqdm==4.67.3
transformers==4.57.6
typer==0.24.1
types-Flask-Cors==6.0.0.20250809
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
wrapt==1.17.3
zipp==3.23.0
